requirements.txt
# DJANGO
asgiref==3.5.2
Django==4.1.1
sqlparse==0.4.3

# DRF
djangorestframework==3.12.4
PyJWT==2.3.0
pytz==2021.3